refactor(myhpi): upper-/lowercase renaming
